requirements.txt
tqdm==4.64.1
requests==2.28.1
numpy==1.24.1
opencv-python==4.7.0.68
argparse=1.4.0
python-dotenv==1.0.0